# Inspect tasks instead of assuming call signatures and return types