@cboulay 2024-05-10
Code looks good. I haven't tested it but I'm on the same arch as you so I don't think it'd help much ;)